refactor(DTO): 잘못 쓰이고 있던 용어 수정
