+++ src/constants/OPTIONS.ts
@@ -1,5 +1,6 @@
 const OPTIONS = {
   movieItemCount: 20,
+  maxPage: 500,
 };
 
 export default OPTIONS;
