+++ src/interfaces/FetchMovieListDTO.ts
