+++ src/interfaces/IResponseDTO.ts
@@ -1,0 +1,7 @@
+// interface IResponse {
+//   status_code: number;
+//   status_message: string;
+//   success: boolean;
+// }
+
+// export default IResponse;
